backend/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
mysql-connector-python==8.3.0
requests==2.31.0
python-dotenv==1.0.0